updat
This application will enable the hooks in the developer application by downloading the source code template from the git repository; inject the code in the application to call those hooks in an orderly fashion. Prehooks will be called before an event, onhook will be called during the event and post hook will be called post the event. The sequence will be called on successful response from the previous event. If there are any errors, the subsequent hooks will not be called.